[IMP] Added tests_rest_permissions
